fix: shopify user create

+++ database/migrations/2024_08_14_054242_create_shopify_users_table.php
@@ -14,7 +14,7 @@ return new class () extends Migration {
         Schema::create('shopify_users', function (Blueprint $table) {
             $table->increments('id');
             $table=$this->groupOrgRelationship($table);
-            $table->unsignedSmallInteger('customer_id')->index();
+            $table->unsignedInteger('customer_id')->index();
             $table->foreign('customer_id')->references('id')->on('customers');
             $table->string('slug')->unique()->collation('und_ns');
             $table->boolean('status')->default(true)->index();
